Move EventCreationView to views folder for clarity and import it in googlecalendar.py
